requirements.txt
# Core data
pandas>=2.1,<3.1
numpy>=1.24,<2.0
scipy>=1.11
pyarrow>=14.0

# ML models
scikit-learn>=1.3,<1.7
xgboost>=2.0
lightgbm>=4.0
catboost>=1.2

# Hyperparameter tuning
optuna>=3.4

# Interpretability
shap>=0.42

# Visualization
matplotlib>=3.7
seaborn>=0.12

# PDF report
fpdf2>=2.7

# Time series
statsmodels>=0.14

# Utilities
tqdm>=4.65
pyyaml>=6.0
joblib>=1.3
requests>=2.31

# Testing
pytest>=7.4
